Fix for HoldFor
I found that HoldFor was not working with multiple layers. Found the error in a leftover `return` from the previous update.
Found an issue when HoldFor was called with single selected keyframe and `affect non selected`=True, regardless of the number it will always add to the distance.(something is wrong with the direction calculation in this case)
Abstracted the binary integer search to the Tools.py file for Nudge