Fixing module dependency installation
Blender 4.5 is in beta, and since I use nightly builds, I finally went and fixed the imports. Changed the call to a powershell command and added some debug printing if it fails installing the module. Also placed some initial `ifs` for cross platform support...linux and windows. But so far this part works only on Windows